api/requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.1
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
flake8==6.0.0
h11==0.14.0
idna==3.4
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
motor==3.1.2
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyflakes==3.0.1
Pygments==2.15.0
PyJWT==2.7.0
pymongo==4.3.3
python-multipart==0.0.6
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.26.1
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.21.1
wcwidth==0.2.6
